Move calculate priority function to the separate folder, correct post function for task
